# Remove StringBuilder marshaling from X509Certificates