Whittling away the remaining StringBuilder marshaling in corefx...

This removes it from X509Certificates. This avoids allocations for the StringBuilder itself, and if the data is small enough to be put on the stack, also underlying the char[].  It also avoids marshaling overheads associated with doing a wstrlen on the resulting data, as we're told exactly how long the string is.

cc: @bartonjs 

Example:
```
[Benchmark]
public static string FriendlyName() => x509Cert.FriendlyName;

       Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
------------- |---------:|---------:|---------:|-------:|----------:|
 Before       | 218.2 ns | 4.432 ns | 5.276 ns | 0.0379 |     160 B |
 After        | 177.8 ns | 1.460 ns | 1.365 ns | 0.0074 |      32 B |
```